Move more logic to ops from hook
